chore: 아이콘 파일 통합
