--- components/icons/arrow/ArrowLeftIcon.tsx
@@ -1,22 +1,0 @@
-import { SvgComponentProps } from '@/types/svgComponentProps';
-
-export const ArrowLeftIcon = ({ ...props }: SvgComponentProps) => {
-  return (
-    <svg
-      {...props}
-      xmlns='http://www.w3.org/2000/svg'
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-    >
-      <path
-        d='M15 18L9 12L15 6'
-        stroke='#23282E'
-        strokeWidth='2'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
