--- components/icons/arrow/ArrowRightIcon.tsx
@@ -1,9 +1,0 @@
-import { SvgComponentProps } from '@/types/svgComponentProps';
-
-export const ArrowRightIcon = (props: SvgComponentProps) => {
-  return (
-    <svg viewBox='0 0 8 12' xmlns='http://www.w3.org/2000/svg' {...props}>
-      <path d='M1 1L6.5 6.5L1 11.5' strokeLinecap='round' />
-    </svg>
-  );
-};
