--- components/icons/bottom-nav/AssetManagementIcon.tsx
@@ -1,27 +1,0 @@
-export const AssetManagementIcon = () => {
-  return (
-    <svg xmlns='http://www.w3.org/2000/svg' width='21' height='22' viewBox='0 0 21 22' fill='none'>
-      <path
-        d='M10.75 17.8691L8.96504 15.7616C8.17754 14.8316 8.23004 13.4516 9.09254 12.5891C10.0075 11.6741 11.4925 11.6741 12.4075 12.5891C13.27 13.4516 13.3225 14.8316 12.535 15.7616L10.75 17.8691Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M10.75 1.36914C4.75 1.36914 4 2.11914 4 2.86914C4 3.61914 10.75 5.11914 10.75 5.11914C10.75 5.11914 17.5 3.61914 17.5 2.86914C17.5 2.11914 16.75 1.36914 10.75 1.36914Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M15.9475 20.6216C17.83 20.7116 19.1125 20.1566 19.375 19.3691C19.375 19.3691 20.125 14.1191 20.125 11.8691C20.125 7.82664 10.75 5.11914 10.75 5.11914C10.75 5.11914 1.375 7.82664 1.375 11.8691C1.375 14.1191 2.125 19.3691 2.125 19.3691C2.5 20.8691 8.5 21.6191 10.75 17.8691C11.3125 18.8066 12.1225 19.4666 13.03 19.9091'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
