+++ components/icons/bottom-nav/BottomNavIcon.tsx
@@ -1,0 +1,160 @@
+export const AssetManagementIcon = () => {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='21' height='22' viewBox='0 0 21 22' fill='none'>
+      <path
+        d='M10.75 17.8691L8.96504 15.7616C8.17754 14.8316 8.23004 13.4516 9.09254 12.5891C10.0075 11.6741 11.4925 11.6741 12.4075 12.5891C13.27 13.4516 13.3225 14.8316 12.535 15.7616L10.75 17.8691Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M10.75 1.36914C4.75 1.36914 4 2.11914 4 2.86914C4 3.61914 10.75 5.11914 10.75 5.11914C10.75 5.11914 17.5 3.61914 17.5 2.86914C17.5 2.11914 16.75 1.36914 10.75 1.36914Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M15.9475 20.6216C17.83 20.7116 19.1125 20.1566 19.375 19.3691C19.375 19.3691 20.125 14.1191 20.125 11.8691C20.125 7.82664 10.75 5.11914 10.75 5.11914C10.75 5.11914 1.375 7.82664 1.375 11.8691C1.375 14.1191 2.125 19.3691 2.125 19.3691C2.5 20.8691 8.5 21.6191 10.75 17.8691C11.3125 18.8066 12.1225 19.4666 13.03 19.9091'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+    </svg>
+  );
+};
+
+export const BudgetCalendarIcon = () => {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='21' height='21' viewBox='0 0 21 21' fill='none'>
+      <path
+        d='M19.8619 12.433V18.5C19.8619 19.3284 19.1903 20 18.3619 20H2.63794C1.80951 20 1.13794 19.3284 1.13794 18.5V11.7725C1.13794 10.9441 1.80951 10.2725 2.63794 10.2725H18.3619C19.1903 10.2725 19.8619 9.60094 19.8619 8.77251V7.32369C19.8619 6.82172 19.3769 6.46319 18.8808 6.53962C15.8723 7.00309 14.1007 5.23145 14.1007 5.23145C14.1007 5.23145 13.0204 6.67175 10.4999 6.67175C7.97938 6.67175 6.89915 5.23145 6.89915 5.23145C6.89915 5.23145 5.67086 7.07389 2.07995 6.50945C1.59831 6.43374 1.13794 6.78897 1.13794 7.27653V8.11205'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M6.89925 4.87694L5.92705 3.7319C5.49495 3.2278 5.53096 2.47164 5.99906 2.00354C6.49597 1.50663 7.30254 1.50663 7.79944 2.00354C8.26754 2.47164 8.29635 3.2278 7.87146 3.7319L6.89925 4.87694Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M14.1007 4.87694L13.1285 3.7319C12.6964 3.2278 12.7324 2.47164 13.2005 2.00354C13.6974 1.50663 14.504 1.50663 15.0009 2.00354C15.469 2.47164 15.4978 3.2278 15.0729 3.7319L14.1007 4.87694Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+    </svg>
+  );
+};
+
+export const FinancialProductIcon = () => {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='23' height='20' viewBox='0 0 23 20' fill='none'>
+      <g clipPath='url(#clip0_1651_17724)'>
+        <path
+          d='M6.75014 5.6725L5.34764 4.0225C4.73264 3.295 4.77014 2.215 5.45264 1.54C6.17264 0.82 7.32764 0.82 8.04764 1.54C8.72264 2.215 8.76764 3.295 8.15264 4.0225L6.75014 5.6725Z'
+          stroke='currentColor'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+        <path
+          d='M15.7501 5.6725L14.3476 4.0225C13.7326 3.295 13.7701 2.215 14.4526 1.54C15.1726 0.82 16.3276 0.82 17.0476 1.54C17.7226 2.215 17.7676 3.295 17.1526 4.0225L15.7501 5.6725Z'
+          stroke='currentColor'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+        <path
+          d='M17.8157 17.5855C17.4741 18.4397 16.6467 18.9999 15.7267 18.9999H6.77332C5.85329 18.9999 5.02594 18.4397 4.68425 17.5855L1.57284 9.80697C1.17872 8.82167 1.91111 7.75245 2.95504 7.56177C5.7551 7.05031 6.75 5.67236 6.75 5.67236C6.75 5.67236 8.25 6.99988 11.25 6.99988C14.25 6.99988 15.75 5.67236 15.75 5.67236C15.75 5.67236 16.7449 7.05031 19.545 7.56177C20.5889 7.75245 21.3213 8.82167 20.9272 9.80696L17.8157 17.5855Z'
+          stroke='currentColor'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+        <path
+          d='M6.75 11.6724L7.5 14.6724'
+          stroke='#B8BEC4'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+        <path
+          d='M11.2764 11.627L11.2244 14.7188'
+          stroke='#B8BEC4'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+        <path
+          d='M15.75 11.6724L15 14.6724'
+          stroke='#B8BEC4'
+          strokeWidth='1.5'
+          strokeLinecap='round'
+          strokeLinejoin='round'
+        />
+      </g>
+      <defs>
+        <clipPath id='clip0_1651_17724'>
+          <rect width='22.5' height='19.6725' fill='white' transform='translate(0 0.25)' />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
+
+export const HomeIcon = () => {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='22' height='21' viewBox='0 0 22 21' fill='none'>
+      <path
+        d='M11 16.9448L9.21504 14.8373C8.42754 13.9073 8.48004 12.5273 9.34254 11.6648C10.2575 10.7498 11.7425 10.7498 12.6575 11.6648C13.52 12.5273 13.5725 13.9073 12.785 14.8373L11 16.9448Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M1.25 7.94482L10.0854 1.14834C10.6246 0.733585 11.3754 0.733585 11.9146 1.14834L20.75 7.94482'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M3.5 9.44482V18.2498C3.5 19.0782 4.17506 19.76 4.99857 19.6699C9.18122 19.2122 11 16.9448 11 16.9448C11 16.9448 12.8188 19.2122 17.0014 19.6699C17.8249 19.76 18.5 19.0782 18.5 18.2498V6.44482'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+    </svg>
+  );
+};
+
+export const NavAllIcon = () => {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='20' height='20' viewBox='0 0 20 20' fill='none'>
+      <path
+        d='M5.15364 6.54051L3.50595 4.59513C2.77903 3.73666 2.82749 2.46282 3.62364 1.66666C4.46826 0.822049 5.83903 0.822049 6.68364 1.66666C7.4798 2.46282 7.52826 3.73666 6.80133 4.59513L5.15364 6.54051Z'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M1 8.61734C3.76923 8.61734 5.15385 6.54041 5.15385 6.54041C5.15385 6.54041 7.23077 8.61734 11.3846 8.61734H17.5C18.3284 8.61734 19 7.94577 19 7.11734V2.50195C19 1.67353 18.3284 1.00195 17.5 1.00195H12.8846C12.0562 1.00195 11.3846 1.67353 11.3846 2.50195V17.502C11.3846 18.3304 12.0562 19.002 12.8846 19.002H17.5C18.3284 19.002 19 18.3304 19 17.502V12.8866C19 12.0581 18.3284 11.3866 17.5 11.3866H2.5C1.67157 11.3866 1 12.0581 1 12.8866V17.5353C1 18.3669 1.67657 19.0398 2.50818 19.0353L7.12357 19.0101C7.94879 19.0056 8.61538 18.3353 8.61538 17.5101V13.4635'
+        stroke='currentColor'
+        strokeWidth='1.5'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+    </svg>
+  );
+};
