--- components/icons/bottom-nav/BudgetCalendarIcon.tsx
@@ -1,27 +1,0 @@
-export const BudgetCalendarIcon = () => {
-  return (
-    <svg xmlns='http://www.w3.org/2000/svg' width='21' height='21' viewBox='0 0 21 21' fill='none'>
-      <path
-        d='M19.8619 12.433V18.5C19.8619 19.3284 19.1903 20 18.3619 20H2.63794C1.80951 20 1.13794 19.3284 1.13794 18.5V11.7725C1.13794 10.9441 1.80951 10.2725 2.63794 10.2725H18.3619C19.1903 10.2725 19.8619 9.60094 19.8619 8.77251V7.32369C19.8619 6.82172 19.3769 6.46319 18.8808 6.53962C15.8723 7.00309 14.1007 5.23145 14.1007 5.23145C14.1007 5.23145 13.0204 6.67175 10.4999 6.67175C7.97938 6.67175 6.89915 5.23145 6.89915 5.23145C6.89915 5.23145 5.67086 7.07389 2.07995 6.50945C1.59831 6.43374 1.13794 6.78897 1.13794 7.27653V8.11205'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M6.89925 4.87694L5.92705 3.7319C5.49495 3.2278 5.53096 2.47164 5.99906 2.00354C6.49597 1.50663 7.30254 1.50663 7.79944 2.00354C8.26754 2.47164 8.29635 3.2278 7.87146 3.7319L6.89925 4.87694Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M14.1007 4.87694L13.1285 3.7319C12.6964 3.2278 12.7324 2.47164 13.2005 2.00354C13.6974 1.50663 14.504 1.50663 15.0009 2.00354C15.469 2.47164 15.4978 3.2278 15.0729 3.7319L14.1007 4.87694Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
