--- components/icons/bottom-nav/FinancialProductIcon.tsx
@@ -1,55 +1,0 @@
-export const FinancialProductIcon = () => {
-  return (
-    <svg xmlns='http://www.w3.org/2000/svg' width='23' height='20' viewBox='0 0 23 20' fill='none'>
-      <g clipPath='url(#clip0_1651_17724)'>
-        <path
-          d='M6.75014 5.6725L5.34764 4.0225C4.73264 3.295 4.77014 2.215 5.45264 1.54C6.17264 0.82 7.32764 0.82 8.04764 1.54C8.72264 2.215 8.76764 3.295 8.15264 4.0225L6.75014 5.6725Z'
-          stroke='currentColor'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-        <path
-          d='M15.7501 5.6725L14.3476 4.0225C13.7326 3.295 13.7701 2.215 14.4526 1.54C15.1726 0.82 16.3276 0.82 17.0476 1.54C17.7226 2.215 17.7676 3.295 17.1526 4.0225L15.7501 5.6725Z'
-          stroke='currentColor'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-        <path
-          d='M17.8157 17.5855C17.4741 18.4397 16.6467 18.9999 15.7267 18.9999H6.77332C5.85329 18.9999 5.02594 18.4397 4.68425 17.5855L1.57284 9.80697C1.17872 8.82167 1.91111 7.75245 2.95504 7.56177C5.7551 7.05031 6.75 5.67236 6.75 5.67236C6.75 5.67236 8.25 6.99988 11.25 6.99988C14.25 6.99988 15.75 5.67236 15.75 5.67236C15.75 5.67236 16.7449 7.05031 19.545 7.56177C20.5889 7.75245 21.3213 8.82167 20.9272 9.80696L17.8157 17.5855Z'
-          stroke='currentColor'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-        <path
-          d='M6.75 11.6724L7.5 14.6724'
-          stroke='#B8BEC4'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-        <path
-          d='M11.2764 11.627L11.2244 14.7188'
-          stroke='#B8BEC4'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-        <path
-          d='M15.75 11.6724L15 14.6724'
-          stroke='#B8BEC4'
-          strokeWidth='1.5'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-      </g>
-      <defs>
-        <clipPath id='clip0_1651_17724'>
-          <rect width='22.5' height='19.6725' fill='white' transform='translate(0 0.25)' />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
