--- components/icons/bottom-nav/HomeIcon.tsx
@@ -1,27 +1,0 @@
-export const HomeIcon = () => {
-  return (
-    <svg xmlns='http://www.w3.org/2000/svg' width='22' height='21' viewBox='0 0 22 21' fill='none'>
-      <path
-        d='M11 16.9448L9.21504 14.8373C8.42754 13.9073 8.48004 12.5273 9.34254 11.6648C10.2575 10.7498 11.7425 10.7498 12.6575 11.6648C13.52 12.5273 13.5725 13.9073 12.785 14.8373L11 16.9448Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M1.25 7.94482L10.0854 1.14834C10.6246 0.733585 11.3754 0.733585 11.9146 1.14834L20.75 7.94482'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M3.5 9.44482V18.2498C3.5 19.0782 4.17506 19.76 4.99857 19.6699C9.18122 19.2122 11 16.9448 11 16.9448C11 16.9448 12.8188 19.2122 17.0014 19.6699C17.8249 19.76 18.5 19.0782 18.5 18.2498V6.44482'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
