--- components/icons/bottom-nav/NavAllIcon.tsx
@@ -1,20 +1,0 @@
-export const NavAllIcon = () => {
-  return (
-    <svg xmlns='http://www.w3.org/2000/svg' width='20' height='20' viewBox='0 0 20 20' fill='none'>
-      <path
-        d='M5.15364 6.54051L3.50595 4.59513C2.77903 3.73666 2.82749 2.46282 3.62364 1.66666C4.46826 0.822049 5.83903 0.822049 6.68364 1.66666C7.4798 2.46282 7.52826 3.73666 6.80133 4.59513L5.15364 6.54051Z'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M1 8.61734C3.76923 8.61734 5.15385 6.54041 5.15385 6.54041C5.15385 6.54041 7.23077 8.61734 11.3846 8.61734H17.5C18.3284 8.61734 19 7.94577 19 7.11734V2.50195C19 1.67353 18.3284 1.00195 17.5 1.00195H12.8846C12.0562 1.00195 11.3846 1.67353 11.3846 2.50195V17.502C11.3846 18.3304 12.0562 19.002 12.8846 19.002H17.5C18.3284 19.002 19 18.3304 19 17.502V12.8866C19 12.0581 18.3284 11.3866 17.5 11.3866H2.5C1.67157 11.3866 1 12.0581 1 12.8866V17.5353C1 18.3669 1.67657 19.0398 2.50818 19.0353L7.12357 19.0101C7.94879 19.0056 8.61538 18.3353 8.61538 17.5101V13.4635'
-        stroke='currentColor'
-        strokeWidth='1.5'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
