+++ components/icons/index.ts
@@ -1,7 +1,4 @@
-export * from './arrow/ArrowRightIcon';
-export * from './arrow/ArrowUpFillIcon';
-export * from './arrow/ArrowLeftIcon';
-export * from './arrow/ArrowBackIcon';
+export * from './arrow/ArrowIcon';
 export * from './BusIcon';
 export * from './CheckIcon';
 export * from './CreditCardIcon';
@@ -9,8 +6,5 @@ export * from './KbCardIcon';
 export * from './MonthlyRentIcon';
 export * from './ShinaHanCardIcon';
 export * from './SubsServiceIcon';
-export * from './bottom-nav/AssetManagementIcon';
-export * from './bottom-nav/BudgetCalendarIcon';
-export * from './bottom-nav/FinancialProductIcon';
-export * from './bottom-nav/HomeIcon';
-export * from './bottom-nav/NavAllIcon';
+export * from './bottom-nav/BottomNavIcon';
+export * from './weather/WeatherIcon';
