--- components/icons/weather/WeatherFive.tsx
@@ -1,69 +1,0 @@
-import React from 'react';
-
-const WeatherFive = () => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='68'
-      height='100'
-      viewBox='0 0 68 100'
-      fill='none'
-    >
-      <path
-        d='M36 0.00292969H32C14.33 0.00292969 0 14.3329 0 32.0029V88.0029C0 94.6329 5.37 100.003 12 100.003H56C62.63 100.003 68 94.6329 68 88.0029V32.0029C68 14.3329 53.67 0.00292969 36 0.00292969Z'
-        fill='#747A81'
-      />
-      <path
-        d='M53.44 79.6222C27.75 82.8923 30.44 66.7123 21.96 67.6923C18.71 68.0623 17.32 72.2822 15.69 69.1822C7.96 54.4922 3.32 59.1222 0 67.7422V89.0822H64.12V79.4722L53.44 79.6222Z'
-        fill='#53A581'
-      />
-      <path
-        d='M41.95 73.2139C28.23 77.2939 14.51 82.1639 0 81.5039V88.0039C0 94.6339 5.37 100.004 12 100.004H56C62.63 100.004 68 94.6339 68 88.0039V70.3639C59.34 69.1839 50.42 70.8539 41.95 73.2139Z'
-        fill='#6CC496'
-      />
-      <path
-        d='M44.8097 25.8121C43.6597 25.7221 42.5397 25.8221 41.4897 26.1021C40.0197 26.4821 38.4797 25.9121 37.5597 24.7121C35.0897 21.4721 31.1497 19.4221 26.7397 19.5421C19.8897 19.7321 14.2697 25.2921 13.9997 32.1521C13.7197 39.6321 19.6997 45.7921 27.1197 45.7921H43.9997C49.7497 45.7921 54.3697 40.9421 53.9797 35.1021C53.6497 30.2021 49.7097 26.2121 44.8197 25.8221L44.8097 25.8121Z'
-        fill='#B8BEC4'
-      />
-      <path
-        d='M28.9297 34.6923C27.6797 33.7523 26.4297 33.7523 25.1797 34.6923'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M21.1897 32.0321C21.1897 32.7221 21.7497 33.2821 22.4397 33.2821C23.1297 33.2821 23.6897 32.7221 23.6897 32.0321C23.6897 31.9721 23.6897 31.9121 23.6697 31.8521L21.3797 31.3721C21.2597 31.5621 21.1897 31.7921 21.1897 32.0321Z'
-        fill='#23282E'
-      />
-      <path
-        d='M33.0597 32.0321C33.0597 32.7221 32.4997 33.2821 31.8097 33.2821C31.1197 33.2821 30.5597 32.7221 30.5597 32.0321C30.5597 31.9721 30.5597 31.9121 30.5797 31.8521L32.8697 31.3721C32.9897 31.5621 33.0597 31.7921 33.0597 32.0321Z'
-        fill='#23282E'
-      />
-      <path
-        d='M31.5497 59.2225L37.8897 51.0325C38.2897 50.5125 37.9197 49.7525 37.2597 49.7525H35.3097L39.7597 42.2225C40.0697 41.6925 39.6897 41.0225 39.0797 41.0225H32.2798C31.9498 41.0225 31.6597 41.2225 31.5397 41.5325L28.1797 50.2625C27.9797 50.7825 28.3597 51.3425 28.9197 51.3425H31.7897L30.1497 58.5625C29.9597 59.3925 31.0297 59.8925 31.5497 59.2225Z'
-        fill='#FFD34C'
-      />
-      <path
-        d='M21.2996 48.4521C20.3596 49.3921 18.6497 51.8521 18.6497 53.7421C18.6497 55.2021 19.8296 56.3921 21.2996 56.3921C22.7696 56.3921 23.9496 55.2121 23.9496 53.7421C23.9496 51.8521 22.2496 49.3921 21.2996 48.4521Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M46.6996 48.4521C45.7596 49.3921 44.0496 51.8521 44.0496 53.7421C44.0496 55.2021 45.2296 56.3921 46.6996 56.3921C48.1696 56.3921 49.3495 55.2121 49.3495 53.7421C49.3495 51.8521 47.6496 49.3921 46.6996 48.4521Z'
-        fill='#B6F2FF'
-      />
-      <g opacity='0.1'>
-        <path
-          d='M58.41 18.3131C53.63 22.5731 62.49 23.2231 66.7 22.9731C65.45 18.7331 63.36 14.8531 60.6 11.5431C53.02 11.2731 46.04 17.1931 58.41 18.3131Z'
-          fill='white'
-        />
-        <path
-          d='M0.23999 28.1729C0.0899902 29.4329 0 30.7029 0 32.0029V35.9129C9 35.2429 9.82999 28.6729 0.23999 28.1829V28.1729Z'
-          fill='white'
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default WeatherFive;
