+++ components/icons/weather/WeatherIcon.tsx
@@ -1,0 +1,350 @@
+import React from 'react';
+
+export const WeatherZero = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      width='68'
+      height='100'
+      viewBox='0 0 68 100'
+      fill='none'
+    >
+      <path
+        d='M36 0H32C14.33 0 0 14.33 0 32V88C0 94.63 5.37 100 12 100H56C62.63 100 68 94.63 68 88V32C68 14.33 53.67 0 36 0Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M53.44 79.6193C27.75 82.8893 30.44 66.7093 21.96 67.6893C18.71 68.0593 17.32 72.2793 15.69 69.1793C7.96 54.4893 3.32 59.1193 0 67.7393V89.0793H64.12V79.4693L53.44 79.6193Z'
+        fill='#55B58E'
+      />
+      <path
+        d='M41.95 73.2109C28.23 77.2909 14.51 82.1609 0 81.5009V88.0009C0 94.6309 5.37 100.001 12 100.001H56C62.63 100.001 68 94.6309 68 88.0009V70.3609C59.34 69.1809 50.42 70.8509 41.95 73.2109Z'
+        fill='#7AD7A9'
+      />
+      <g opacity='0.3'>
+        <path
+          d='M58.41 18.3102C53.63 22.5702 62.49 23.2202 66.7 22.9702C65.45 18.7302 63.36 14.8502 60.6 11.5402C53.02 11.2702 46.04 17.1902 58.41 18.3102Z'
+          fill='white'
+        />
+        <path
+          d='M0.24 28.1699C0.0899998 29.4299 0 30.6999 0 31.9999V35.9099C9 35.2399 9.83 28.6699 0.24 28.1799V28.1699Z'
+          fill='white'
+        />
+      </g>
+      <path
+        d='M33.9996 53.1211C43.6646 53.1211 51.4996 45.2861 51.4996 35.6211C51.4996 25.9561 43.6646 18.1211 33.9996 18.1211C24.3347 18.1211 16.4996 25.9561 16.4996 35.6211C16.4996 45.2861 24.3347 53.1211 33.9996 53.1211Z'
+        fill='#FF5C46'
+      />
+      <path
+        d='M33.9996 21.8711C26.4096 21.8711 20.2496 28.0311 20.2496 35.6211C20.2496 43.2111 26.4096 49.3711 33.9996 49.3711C41.5896 49.3711 47.7496 43.2111 47.7496 35.6211C47.7496 28.0311 41.5896 21.8711 33.9996 21.8711ZM33.9996 45.0011C28.8196 45.0011 24.6196 40.8011 24.6196 35.6211C24.6196 30.4411 28.8196 26.2411 33.9996 26.2411C39.1796 26.2411 43.3796 30.4411 43.3796 35.6211C43.3796 40.8011 39.1796 45.0011 33.9996 45.0011Z'
+        fill='#FFD34C'
+      />
+      <path
+        d='M33.9996 26.251C28.8196 26.251 24.6196 30.451 24.6196 35.631C24.6196 40.811 28.8196 45.011 33.9996 45.011C39.1796 45.011 43.3796 40.811 43.3796 35.631C43.3796 30.451 39.1796 26.251 33.9996 26.251ZM33.9996 40.621C31.2396 40.621 28.9996 38.381 28.9996 35.621C28.9996 32.861 31.2396 30.621 33.9996 30.621C36.7596 30.621 38.9996 32.861 38.9996 35.621C38.9996 38.381 36.7596 40.621 33.9996 40.621Z'
+        fill='#7AD7A9'
+      />
+      <path
+        d='M33.9996 40.6211C36.7611 40.6211 38.9996 38.3825 38.9996 35.6211C38.9996 32.8597 36.7611 30.6211 33.9996 30.6211C31.2382 30.6211 28.9996 32.8597 28.9996 35.6211C28.9996 38.3825 31.2382 40.6211 33.9996 40.6211Z'
+        fill='#5A7EFF'
+      />
+      <path
+        d='M44.8097 38.1607C43.6597 38.0707 42.5397 38.1707 41.4897 38.4507C40.0197 38.8307 38.4797 38.2607 37.5597 37.0607C35.0897 33.8207 31.1497 31.7707 26.7397 31.8907C19.8897 32.0807 14.2597 37.6407 13.9997 44.5007C13.7197 51.9807 19.6997 58.1407 27.1197 58.1407H43.9997C49.7497 58.1407 54.3697 53.2907 53.9797 47.4507C53.6497 42.5507 49.7097 38.5607 44.8197 38.1707L44.8097 38.1607Z'
+        fill='white'
+      />
+      <path
+        d='M25.2496 47.501C26.4996 48.441 27.7496 48.441 28.9996 47.501'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M21.3297 43.1211L23.3697 45.0011L21.3297 46.8711'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M20.5597 45.001H23.3697'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M32.9196 43.1211L30.8696 45.0011L32.9196 46.8711'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+      <path
+        d='M33.6896 45.001H30.8696'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeLinecap='round'
+        strokeLinejoin='round'
+      />
+    </svg>
+  );
+};
+
+export const WeatherOne = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      width='68'
+      height='100'
+      viewBox='0 0 68 100'
+      fill='none'
+    >
+      <path
+        d='M36 0H32C14.33 0 0 14.33 0 32V88C0 94.63 5.37 100 12 100H56C62.63 100 68 94.63 68 88V32C68 14.33 53.67 0 36 0Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M53.44 79.6193C27.75 82.8893 30.44 66.7093 21.96 67.6893C18.71 68.0593 17.32 72.2793 15.69 69.1793C7.96 54.4893 3.32 59.1193 0 67.7393V89.0793H64.12V79.4693L53.44 79.6193Z'
+        fill='#55B58E'
+      />
+      <path
+        d='M41.95 73.2109C28.23 77.2909 14.51 82.1609 0 81.5009V88.0009C0 94.6309 5.37 100.001 12 100.001H56C62.63 100.001 68 94.6309 68 88.0009V70.3609C59.34 69.1809 50.42 70.8509 41.95 73.2109Z'
+        fill='#7AD7A9'
+      />
+      <path
+        d='M39.7001 26.6293L47.5501 25.2393C47.9801 25.1593 48.3501 25.5393 48.2701 25.9593L46.8801 33.8093C46.8401 34.0493 46.9401 34.2893 47.1401 34.4293L53.6701 38.9993C54.0301 39.2493 54.0301 39.7793 53.6701 40.0193L47.1401 44.5893C46.9401 44.7293 46.8401 44.9693 46.8801 45.2093L48.2701 53.0593C48.3501 53.4893 47.9701 53.8593 47.5501 53.7793L39.7001 52.3893C39.4601 52.3493 39.2201 52.4493 39.0801 52.6493L34.5101 59.1793C34.2601 59.5393 33.7301 59.5393 33.4901 59.1793L28.9201 52.6493C28.7801 52.4493 28.5401 52.3493 28.3001 52.3893L20.4501 53.7793C20.0201 53.8593 19.6501 53.4793 19.7301 53.0593L21.1201 45.2093C21.1601 44.9693 21.0601 44.7293 20.8601 44.5893L14.3301 40.0193C13.9701 39.7693 13.9701 39.2393 14.3301 38.9993L20.8601 34.4293C21.0601 34.2893 21.1601 34.0493 21.1201 33.8093L19.7301 25.9593C19.6501 25.5293 20.0301 25.1593 20.4501 25.2393L28.3001 26.6293C28.5401 26.6693 28.7801 26.5693 28.9201 26.3693L33.4901 19.8393C33.7401 19.4793 34.2701 19.4793 34.5101 19.8393L39.0801 26.3693C39.2201 26.5693 39.4601 26.6693 39.7001 26.6293Z'
+        fill='#FFD34C'
+      />
+      <path
+        d='M29.3101 40.7793C30.0004 40.7793 30.5601 40.2197 30.5601 39.5293C30.5601 38.8389 30.0004 38.2793 29.3101 38.2793C28.6197 38.2793 28.0601 38.8389 28.0601 39.5293C28.0601 40.2197 28.6197 40.7793 29.3101 40.7793Z'
+        fill='#23282E'
+      />
+      <path
+        d='M38.6901 40.7793C39.3804 40.7793 39.9401 40.2197 39.9401 39.5293C39.9401 38.8389 39.3804 38.2793 38.6901 38.2793C37.9997 38.2793 37.4401 38.8389 37.4401 39.5293C37.4401 40.2197 37.9997 40.7793 38.6901 40.7793Z'
+        fill='#23282E'
+      />
+      <path
+        d='M32.12 42.0293C33.37 42.9693 34.62 42.9693 35.87 42.0293'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <g opacity='0.3'>
+        <path
+          d='M58.41 18.3102C53.63 22.5702 62.49 23.2202 66.7 22.9702C65.45 18.7302 63.36 14.8502 60.6 11.5402C53.02 11.2702 46.04 17.1902 58.41 18.3102Z'
+          fill='white'
+        />
+        <path
+          d='M0.240005 28.1699C0.0900055 29.4299 0 30.6999 0 31.9999V35.9099C9 35.2399 9.83001 28.6699 0.240005 28.1799V28.1699Z'
+          fill='white'
+        />
+      </g>
+    </svg>
+  );
+};
+
+export const WeatherTwo = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      width='68'
+      height='100'
+      viewBox='0 0 68 100'
+      fill='none'
+    >
+      <path
+        d='M36 0.000976562H32C14.33 0.000976562 0 14.331 0 32.001V88.001C0 94.631 5.37 100.001 12 100.001H56C62.63 100.001 68 94.631 68 88.001V32.001C68 14.331 53.67 0.000976562 36 0.000976562Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M53.44 79.6203C27.75 82.8903 30.44 66.7103 21.96 67.6903C18.71 68.0603 17.32 72.2803 15.69 69.1803C7.96 54.4903 3.32 59.1203 0 67.7403V89.0803H64.12V79.4703L53.44 79.6203Z'
+        fill='#55B58E'
+      />
+      <path
+        d='M41.95 73.2119C28.23 77.2919 14.51 82.1619 0 81.5019V88.0019C0 94.6319 5.37 100.002 12 100.002H56C62.63 100.002 68 94.6319 68 88.0019V70.3619C59.34 69.1819 50.42 70.8519 41.95 73.2119Z'
+        fill='#7AD7A9'
+      />
+      <path
+        d='M43.4497 24.7902L48.5497 23.8902C49.1897 23.7802 49.7497 24.3402 49.6397 24.9802L48.7397 30.0802C48.6797 30.4402 48.8297 30.8002 49.1297 31.0102L53.3697 33.9802C53.8997 34.3502 53.8997 35.1402 53.3697 35.5202L49.1297 38.4902C48.8297 38.7002 48.6797 39.0602 48.7397 39.4202L49.6397 44.5202C49.7497 45.1602 49.1897 45.7202 48.5497 45.6102L43.4497 44.7102C43.0897 44.6502 42.7297 44.8002 42.5197 45.1002L39.5497 49.3402C39.1797 49.8702 38.3897 49.8702 38.0097 49.3402L35.0397 45.1002C34.8297 44.8002 34.4697 44.6502 34.1097 44.7102L29.0097 45.6102C28.3697 45.7202 27.8097 45.1602 27.9197 44.5202L28.8197 39.4202C28.8797 39.0602 28.7297 38.7002 28.4297 38.4902L24.1897 35.5202C23.6597 35.1502 23.6597 34.3602 24.1897 33.9802L28.4297 31.0102C28.7297 30.8002 28.8797 30.4402 28.8197 30.0802L27.9197 24.9802C27.8097 24.3402 28.3697 23.7802 29.0097 23.8902L34.1097 24.7902C34.4697 24.8502 34.8297 24.7002 35.0397 24.4002L38.0097 20.1602C38.3797 19.6302 39.1697 19.6302 39.5497 20.1602L42.5197 24.4002C42.7297 24.7002 43.0897 24.8502 43.4497 24.7902Z'
+        fill='#FFD34C'
+      />
+      <path
+        d='M44.8097 39.5602C43.6597 39.4702 42.5397 39.5702 41.4897 39.8502C40.0197 40.2302 38.4797 39.6602 37.5597 38.4602C35.0897 35.2202 31.1497 33.1702 26.7397 33.2902C19.8897 33.4802 14.2697 39.0402 13.9997 45.9002C13.7197 53.3802 19.6997 59.5402 27.1197 59.5402H43.9997C49.7497 59.5402 54.3697 54.6902 53.9797 48.8502C53.6497 43.9502 49.7097 39.9602 44.8197 39.5702L44.8097 39.5602Z'
+        fill='white'
+      />
+      <path
+        d='M35.2498 33.9004C35.9401 33.9004 36.4998 33.3407 36.4998 32.6504C36.4998 31.96 35.9401 31.4004 35.2498 31.4004C34.5594 31.4004 33.9998 31.96 33.9998 32.6504C33.9998 33.3407 34.5594 33.9004 35.2498 33.9004Z'
+        fill='#23282E'
+      />
+      <path
+        d='M42.1198 33.9004C42.8101 33.9004 43.3698 33.3407 43.3698 32.6504C43.3698 31.96 42.8101 31.4004 42.1198 31.4004C41.4294 31.4004 40.8698 31.96 40.8698 32.6504C40.8698 33.3407 41.4294 33.9004 42.1198 33.9004Z'
+        fill='#23282E'
+      />
+      <path
+        d='M22.4397 47.6504C23.1301 47.6504 23.6897 47.0907 23.6897 46.4004C23.6897 45.71 23.1301 45.1504 22.4397 45.1504C21.7493 45.1504 21.1897 45.71 21.1897 46.4004C21.1897 47.0907 21.7493 47.6504 22.4397 47.6504Z'
+        fill='#23282E'
+      />
+      <path
+        d='M31.8097 47.6504C32.5 47.6504 33.0597 47.0907 33.0597 46.4004C33.0597 45.71 32.5 45.1504 31.8097 45.1504C31.1193 45.1504 30.5597 45.71 30.5597 46.4004C30.5597 47.0907 31.1193 47.6504 31.8097 47.6504Z'
+        fill='#23282E'
+      />
+      <path
+        d='M25.2498 48.9004H28.9998'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M36.8097 35.1504H40.5597'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <g opacity='0.3'>
+        <path
+          d='M58.41 18.3112C53.63 22.5712 62.49 23.2212 66.7 22.9712C65.45 18.7312 63.36 14.8512 60.6 11.5412C53.02 11.2712 46.04 17.1912 58.41 18.3112Z'
+          fill='white'
+        />
+        <path
+          d='M0.23999 28.1709C0.0899902 29.4309 0 30.7009 0 32.0009V35.9109C9 35.2409 9.82999 28.6709 0.23999 28.1809V28.1709Z'
+          fill='white'
+        />
+      </g>
+    </svg>
+  );
+};
+
+export const WeatherThree = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      width='68'
+      height='100'
+      viewBox='0 0 68 100'
+      fill='none'
+    >
+      <path
+        d='M36 0.00195312H32C14.33 0.00195312 0 14.332 0 32.002V88.002C0 94.632 5.37 100.002 12 100.002H56C62.63 100.002 68 94.632 68 88.002V32.002C68 14.332 53.67 0.00195312 36 0.00195312Z'
+        fill='#747A81'
+      />
+      <path
+        d='M53.44 79.6213C27.75 82.8913 30.44 66.7113 21.96 67.6913C18.71 68.0613 17.32 72.2813 15.69 69.1813C7.96 54.4913 3.32 59.1213 0 67.7413V89.0813H64.12V79.4713L53.44 79.6213Z'
+        fill='#53A581'
+      />
+      <path
+        d='M41.95 73.2129C28.23 77.2929 14.51 82.1629 0 81.5029V88.0029C0 94.6329 5.37 100.003 12 100.003H56C62.63 100.003 68 94.6329 68 88.0029V70.3629C59.34 69.1829 50.42 70.8529 41.95 73.2129Z'
+        fill='#6CC496'
+      />
+      <path
+        d='M44.8097 25.8111C43.6597 25.7211 42.5397 25.8211 41.4897 26.1011C40.0197 26.4811 38.4797 25.9111 37.5597 24.7111C35.0897 21.4711 31.1497 19.4211 26.7397 19.5411C19.8897 19.7311 14.2697 25.2911 13.9997 32.1511C13.7197 39.6311 19.6997 45.7911 27.1197 45.7911H43.9997C49.7497 45.7911 54.3697 40.9411 53.9797 35.1011C53.6497 30.2011 49.7097 26.2111 44.8197 25.8211L44.8097 25.8111Z'
+        fill='#B8BEC4'
+      />
+      <path
+        d='M22.4397 33.2812C23.1301 33.2812 23.6897 32.7216 23.6897 32.0312C23.6897 31.3409 23.1301 30.7812 22.4397 30.7812C21.7493 30.7812 21.1897 31.3409 21.1897 32.0312C21.1897 32.7216 21.7493 33.2812 22.4397 33.2812Z'
+        fill='#23282E'
+      />
+      <path
+        d='M31.8097 33.2812C32.5 33.2812 33.0597 32.7216 33.0597 32.0312C33.0597 31.3409 32.5 30.7812 31.8097 30.7812C31.1193 30.7812 30.5597 31.3409 30.5597 32.0312C30.5597 32.7216 31.1193 33.2812 31.8097 33.2812Z'
+        fill='#23282E'
+      />
+      <path
+        d='M28.9297 34.6913C27.6797 33.7513 26.4297 33.7513 25.1797 34.6913'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M33.9998 48.2412C32.6998 49.5812 30.3597 53.0812 30.3597 55.7712C30.3597 57.8512 31.9898 59.5312 33.9998 59.5312C36.0098 59.5312 37.6398 57.8512 37.6398 55.7712C37.6398 53.0812 35.2998 49.5912 33.9998 48.2412Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M23.9097 48.2412C23.0397 49.1412 21.4797 51.4712 21.4797 53.2612C21.4797 54.6512 22.5697 55.7712 23.9097 55.7712C25.2497 55.7712 26.3397 54.6512 26.3397 53.2612C26.3397 51.4712 24.7797 49.1412 23.9097 48.2412Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M44.0897 48.2412C43.2197 49.1412 41.6597 51.4712 41.6597 53.2612C41.6597 54.6512 42.7497 55.7712 44.0897 55.7712C45.4297 55.7712 46.5197 54.6512 46.5197 53.2612C46.5197 51.4712 44.9597 49.1412 44.0897 48.2412Z'
+        fill='#B6F2FF'
+      />
+      <g opacity='0.1'>
+        <path
+          d='M58.41 18.3122C53.63 22.5722 62.49 23.2222 66.7 22.9722C65.45 18.7322 63.36 14.8522 60.6 11.5422C53.02 11.2722 46.04 17.1922 58.41 18.3122Z'
+          fill='white'
+        />
+        <path
+          d='M0.23999 28.1719C0.0899902 29.4319 0 30.7019 0 32.0019V35.9119C9 35.2419 9.82999 28.6719 0.23999 28.1819V28.1719Z'
+          fill='white'
+        />
+      </g>
+    </svg>
+  );
+};
+
+export const WeatherFour = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      width='68'
+      height='100'
+      viewBox='0 0 68 100'
+      fill='none'
+    >
+      <path
+        d='M36 0.00292969H32C14.33 0.00292969 0 14.3329 0 32.0029V88.0029C0 94.6329 5.37 100.003 12 100.003H56C62.63 100.003 68 94.6329 68 88.0029V32.0029C68 14.3329 53.67 0.00292969 36 0.00292969Z'
+        fill='#747A81'
+      />
+      <path
+        d='M53.44 79.6222C27.75 82.8923 30.44 66.7123 21.96 67.6923C18.71 68.0623 17.32 72.2822 15.69 69.1822C7.96 54.4922 3.32 59.1222 0 67.7422V89.0822H64.12V79.4722L53.44 79.6222Z'
+        fill='#53A581'
+      />
+      <path
+        d='M41.95 73.2139C28.23 77.2939 14.51 82.1639 0 81.5039V88.0039C0 94.6339 5.37 100.004 12 100.004H56C62.63 100.004 68 94.6339 68 88.0039V70.3639C59.34 69.1839 50.42 70.8539 41.95 73.2139Z'
+        fill='#6CC496'
+      />
+      <path
+        d='M44.8097 25.8121C43.6597 25.7221 42.5397 25.8221 41.4897 26.1021C40.0197 26.4821 38.4797 25.9121 37.5597 24.7121C35.0897 21.4721 31.1497 19.4221 26.7397 19.5421C19.8897 19.7321 14.2697 25.2921 13.9997 32.1521C13.7197 39.6321 19.6997 45.7921 27.1197 45.7921H43.9997C49.7497 45.7921 54.3697 40.9421 53.9797 35.1021C53.6497 30.2021 49.7097 26.2121 44.8197 25.8221L44.8097 25.8121Z'
+        fill='#B8BEC4'
+      />
+      <path
+        d='M28.9297 34.6923C27.6797 33.7523 26.4297 33.7523 25.1797 34.6923'
+        stroke='#23282E'
+        strokeWidth='1.25'
+        strokeMiterlimit='10'
+        strokeLinecap='round'
+      />
+      <path
+        d='M21.1897 32.0321C21.1897 32.7221 21.7497 33.2821 22.4397 33.2821C23.1297 33.2821 23.6897 32.7221 23.6897 32.0321C23.6897 31.9721 23.6897 31.9121 23.6697 31.8521L21.3797 31.3721C21.2597 31.5621 21.1897 31.7921 21.1897 32.0321Z'
+        fill='#23282E'
+      />
+      <path
+        d='M33.0597 32.0321C33.0597 32.7221 32.4997 33.2821 31.8097 33.2821C31.1197 33.2821 30.5597 32.7221 30.5597 32.0321C30.5597 31.9721 30.5597 31.9121 30.5797 31.8521L32.8697 31.3721C32.9897 31.5621 33.0597 31.7921 33.0597 32.0321Z'
+        fill='#23282E'
+      />
+      <path
+        d='M31.5497 59.2225L37.8897 51.0325C38.2897 50.5125 37.9197 49.7525 37.2597 49.7525H35.3097L39.7597 42.2225C40.0697 41.6925 39.6897 41.0225 39.0797 41.0225H32.2798C31.9498 41.0225 31.6597 41.2225 31.5397 41.5325L28.1797 50.2625C27.9797 50.7825 28.3597 51.3425 28.9197 51.3425H31.7897L30.1497 58.5625C29.9597 59.3925 31.0297 59.8925 31.5497 59.2225Z'
+        fill='#FFD34C'
+      />
+      <path
+        d='M21.2996 48.4521C20.3596 49.3921 18.6497 51.8521 18.6497 53.7421C18.6497 55.2021 19.8296 56.3921 21.2996 56.3921C22.7696 56.3921 23.9496 55.2121 23.9496 53.7421C23.9496 51.8521 22.2496 49.3921 21.2996 48.4521Z'
+        fill='#B6F2FF'
+      />
+      <path
+        d='M46.6996 48.4521C45.7596 49.3921 44.0496 51.8521 44.0496 53.7421C44.0496 55.2021 45.2296 56.3921 46.6996 56.3921C48.1696 56.3921 49.3495 55.2121 49.3495 53.7421C49.3495 51.8521 47.6496 49.3921 46.6996 48.4521Z'
+        fill='#B6F2FF'
+      />
+      <g opacity='0.1'>
+        <path
+          d='M58.41 18.3131C53.63 22.5731 62.49 23.2231 66.7 22.9731C65.45 18.7331 63.36 14.8531 60.6 11.5431C53.02 11.2731 46.04 17.1931 58.41 18.3131Z'
+          fill='white'
+        />
+        <path
+          d='M0.23999 28.1729C0.0899902 29.4329 0 30.7029 0 32.0029V35.9129C9 35.2429 9.82999 28.6729 0.23999 28.1829V28.1729Z'
+          fill='white'
+        />
+      </g>
+    </svg>
+  );
+};
