--- components/icons/weather/WeatherOne.tsx
@@ -1,57 +1,0 @@
-import React from 'react';
-
-const WeatherOne = () => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='68'
-      height='100'
-      viewBox='0 0 68 100'
-      fill='none'
-    >
-      <path
-        d='M36 0H32C14.33 0 0 14.33 0 32V88C0 94.63 5.37 100 12 100H56C62.63 100 68 94.63 68 88V32C68 14.33 53.67 0 36 0Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M53.44 79.6193C27.75 82.8893 30.44 66.7093 21.96 67.6893C18.71 68.0593 17.32 72.2793 15.69 69.1793C7.96 54.4893 3.32 59.1193 0 67.7393V89.0793H64.12V79.4693L53.44 79.6193Z'
-        fill='#55B58E'
-      />
-      <path
-        d='M41.95 73.2109C28.23 77.2909 14.51 82.1609 0 81.5009V88.0009C0 94.6309 5.37 100.001 12 100.001H56C62.63 100.001 68 94.6309 68 88.0009V70.3609C59.34 69.1809 50.42 70.8509 41.95 73.2109Z'
-        fill='#7AD7A9'
-      />
-      <path
-        d='M39.7001 26.6293L47.5501 25.2393C47.9801 25.1593 48.3501 25.5393 48.2701 25.9593L46.8801 33.8093C46.8401 34.0493 46.9401 34.2893 47.1401 34.4293L53.6701 38.9993C54.0301 39.2493 54.0301 39.7793 53.6701 40.0193L47.1401 44.5893C46.9401 44.7293 46.8401 44.9693 46.8801 45.2093L48.2701 53.0593C48.3501 53.4893 47.9701 53.8593 47.5501 53.7793L39.7001 52.3893C39.4601 52.3493 39.2201 52.4493 39.0801 52.6493L34.5101 59.1793C34.2601 59.5393 33.7301 59.5393 33.4901 59.1793L28.9201 52.6493C28.7801 52.4493 28.5401 52.3493 28.3001 52.3893L20.4501 53.7793C20.0201 53.8593 19.6501 53.4793 19.7301 53.0593L21.1201 45.2093C21.1601 44.9693 21.0601 44.7293 20.8601 44.5893L14.3301 40.0193C13.9701 39.7693 13.9701 39.2393 14.3301 38.9993L20.8601 34.4293C21.0601 34.2893 21.1601 34.0493 21.1201 33.8093L19.7301 25.9593C19.6501 25.5293 20.0301 25.1593 20.4501 25.2393L28.3001 26.6293C28.5401 26.6693 28.7801 26.5693 28.9201 26.3693L33.4901 19.8393C33.7401 19.4793 34.2701 19.4793 34.5101 19.8393L39.0801 26.3693C39.2201 26.5693 39.4601 26.6693 39.7001 26.6293Z'
-        fill='#FFD34C'
-      />
-      <path
-        d='M29.3101 40.7793C30.0004 40.7793 30.5601 40.2197 30.5601 39.5293C30.5601 38.8389 30.0004 38.2793 29.3101 38.2793C28.6197 38.2793 28.0601 38.8389 28.0601 39.5293C28.0601 40.2197 28.6197 40.7793 29.3101 40.7793Z'
-        fill='#23282E'
-      />
-      <path
-        d='M38.6901 40.7793C39.3804 40.7793 39.9401 40.2197 39.9401 39.5293C39.9401 38.8389 39.3804 38.2793 38.6901 38.2793C37.9997 38.2793 37.4401 38.8389 37.4401 39.5293C37.4401 40.2197 37.9997 40.7793 38.6901 40.7793Z'
-        fill='#23282E'
-      />
-      <path
-        d='M32.12 42.0293C33.37 42.9693 34.62 42.9693 35.87 42.0293'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <g opacity='0.3'>
-        <path
-          d='M58.41 18.3102C53.63 22.5702 62.49 23.2202 66.7 22.9702C65.45 18.7302 63.36 14.8502 60.6 11.5402C53.02 11.2702 46.04 17.1902 58.41 18.3102Z'
-          fill='white'
-        />
-        <path
-          d='M0.240005 28.1699C0.0900055 29.4299 0 30.6999 0 31.9999V35.9099C9 35.2399 9.83001 28.6699 0.240005 28.1799V28.1699Z'
-          fill='white'
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default WeatherOne;
