--- components/icons/weather/WeatherThree.tsx
@@ -1,69 +1,0 @@
-import React from 'react';
-
-const WeatherThree = () => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='68'
-      height='100'
-      viewBox='0 0 68 100'
-      fill='none'
-    >
-      <path
-        d='M36 0.00195312H32C14.33 0.00195312 0 14.332 0 32.002V88.002C0 94.632 5.37 100.002 12 100.002H56C62.63 100.002 68 94.632 68 88.002V32.002C68 14.332 53.67 0.00195312 36 0.00195312Z'
-        fill='#747A81'
-      />
-      <path
-        d='M53.44 79.6213C27.75 82.8913 30.44 66.7113 21.96 67.6913C18.71 68.0613 17.32 72.2813 15.69 69.1813C7.96 54.4913 3.32 59.1213 0 67.7413V89.0813H64.12V79.4713L53.44 79.6213Z'
-        fill='#53A581'
-      />
-      <path
-        d='M41.95 73.2129C28.23 77.2929 14.51 82.1629 0 81.5029V88.0029C0 94.6329 5.37 100.003 12 100.003H56C62.63 100.003 68 94.6329 68 88.0029V70.3629C59.34 69.1829 50.42 70.8529 41.95 73.2129Z'
-        fill='#6CC496'
-      />
-      <path
-        d='M44.8097 25.8111C43.6597 25.7211 42.5397 25.8211 41.4897 26.1011C40.0197 26.4811 38.4797 25.9111 37.5597 24.7111C35.0897 21.4711 31.1497 19.4211 26.7397 19.5411C19.8897 19.7311 14.2697 25.2911 13.9997 32.1511C13.7197 39.6311 19.6997 45.7911 27.1197 45.7911H43.9997C49.7497 45.7911 54.3697 40.9411 53.9797 35.1011C53.6497 30.2011 49.7097 26.2111 44.8197 25.8211L44.8097 25.8111Z'
-        fill='#B8BEC4'
-      />
-      <path
-        d='M22.4397 33.2812C23.1301 33.2812 23.6897 32.7216 23.6897 32.0312C23.6897 31.3409 23.1301 30.7812 22.4397 30.7812C21.7493 30.7812 21.1897 31.3409 21.1897 32.0312C21.1897 32.7216 21.7493 33.2812 22.4397 33.2812Z'
-        fill='#23282E'
-      />
-      <path
-        d='M31.8097 33.2812C32.5 33.2812 33.0597 32.7216 33.0597 32.0312C33.0597 31.3409 32.5 30.7812 31.8097 30.7812C31.1193 30.7812 30.5597 31.3409 30.5597 32.0312C30.5597 32.7216 31.1193 33.2812 31.8097 33.2812Z'
-        fill='#23282E'
-      />
-      <path
-        d='M28.9297 34.6913C27.6797 33.7513 26.4297 33.7513 25.1797 34.6913'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M33.9998 48.2412C32.6998 49.5812 30.3597 53.0812 30.3597 55.7712C30.3597 57.8512 31.9898 59.5312 33.9998 59.5312C36.0098 59.5312 37.6398 57.8512 37.6398 55.7712C37.6398 53.0812 35.2998 49.5912 33.9998 48.2412Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M23.9097 48.2412C23.0397 49.1412 21.4797 51.4712 21.4797 53.2612C21.4797 54.6512 22.5697 55.7712 23.9097 55.7712C25.2497 55.7712 26.3397 54.6512 26.3397 53.2612C26.3397 51.4712 24.7797 49.1412 23.9097 48.2412Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M44.0897 48.2412C43.2197 49.1412 41.6597 51.4712 41.6597 53.2612C41.6597 54.6512 42.7497 55.7712 44.0897 55.7712C45.4297 55.7712 46.5197 54.6512 46.5197 53.2612C46.5197 51.4712 44.9597 49.1412 44.0897 48.2412Z'
-        fill='#B6F2FF'
-      />
-      <g opacity='0.1'>
-        <path
-          d='M58.41 18.3122C53.63 22.5722 62.49 23.2222 66.7 22.9722C65.45 18.7322 63.36 14.8522 60.6 11.5422C53.02 11.2722 46.04 17.1922 58.41 18.3122Z'
-          fill='white'
-        />
-        <path
-          d='M0.23999 28.1719C0.0899902 29.4319 0 30.7019 0 32.0019V35.9119C9 35.2419 9.82999 28.6719 0.23999 28.1819V28.1719Z'
-          fill='white'
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default WeatherThree;
