--- components/icons/weather/WeatherTwo.tsx
@@ -1,76 +1,0 @@
-import React from 'react';
-
-const WeatherTwo = () => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='68'
-      height='100'
-      viewBox='0 0 68 100'
-      fill='none'
-    >
-      <path
-        d='M36 0.000976562H32C14.33 0.000976562 0 14.331 0 32.001V88.001C0 94.631 5.37 100.001 12 100.001H56C62.63 100.001 68 94.631 68 88.001V32.001C68 14.331 53.67 0.000976562 36 0.000976562Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M53.44 79.6203C27.75 82.8903 30.44 66.7103 21.96 67.6903C18.71 68.0603 17.32 72.2803 15.69 69.1803C7.96 54.4903 3.32 59.1203 0 67.7403V89.0803H64.12V79.4703L53.44 79.6203Z'
-        fill='#55B58E'
-      />
-      <path
-        d='M41.95 73.2119C28.23 77.2919 14.51 82.1619 0 81.5019V88.0019C0 94.6319 5.37 100.002 12 100.002H56C62.63 100.002 68 94.6319 68 88.0019V70.3619C59.34 69.1819 50.42 70.8519 41.95 73.2119Z'
-        fill='#7AD7A9'
-      />
-      <path
-        d='M43.4497 24.7902L48.5497 23.8902C49.1897 23.7802 49.7497 24.3402 49.6397 24.9802L48.7397 30.0802C48.6797 30.4402 48.8297 30.8002 49.1297 31.0102L53.3697 33.9802C53.8997 34.3502 53.8997 35.1402 53.3697 35.5202L49.1297 38.4902C48.8297 38.7002 48.6797 39.0602 48.7397 39.4202L49.6397 44.5202C49.7497 45.1602 49.1897 45.7202 48.5497 45.6102L43.4497 44.7102C43.0897 44.6502 42.7297 44.8002 42.5197 45.1002L39.5497 49.3402C39.1797 49.8702 38.3897 49.8702 38.0097 49.3402L35.0397 45.1002C34.8297 44.8002 34.4697 44.6502 34.1097 44.7102L29.0097 45.6102C28.3697 45.7202 27.8097 45.1602 27.9197 44.5202L28.8197 39.4202C28.8797 39.0602 28.7297 38.7002 28.4297 38.4902L24.1897 35.5202C23.6597 35.1502 23.6597 34.3602 24.1897 33.9802L28.4297 31.0102C28.7297 30.8002 28.8797 30.4402 28.8197 30.0802L27.9197 24.9802C27.8097 24.3402 28.3697 23.7802 29.0097 23.8902L34.1097 24.7902C34.4697 24.8502 34.8297 24.7002 35.0397 24.4002L38.0097 20.1602C38.3797 19.6302 39.1697 19.6302 39.5497 20.1602L42.5197 24.4002C42.7297 24.7002 43.0897 24.8502 43.4497 24.7902Z'
-        fill='#FFD34C'
-      />
-      <path
-        d='M44.8097 39.5602C43.6597 39.4702 42.5397 39.5702 41.4897 39.8502C40.0197 40.2302 38.4797 39.6602 37.5597 38.4602C35.0897 35.2202 31.1497 33.1702 26.7397 33.2902C19.8897 33.4802 14.2697 39.0402 13.9997 45.9002C13.7197 53.3802 19.6997 59.5402 27.1197 59.5402H43.9997C49.7497 59.5402 54.3697 54.6902 53.9797 48.8502C53.6497 43.9502 49.7097 39.9602 44.8197 39.5702L44.8097 39.5602Z'
-        fill='white'
-      />
-      <path
-        d='M35.2498 33.9004C35.9401 33.9004 36.4998 33.3407 36.4998 32.6504C36.4998 31.96 35.9401 31.4004 35.2498 31.4004C34.5594 31.4004 33.9998 31.96 33.9998 32.6504C33.9998 33.3407 34.5594 33.9004 35.2498 33.9004Z'
-        fill='#23282E'
-      />
-      <path
-        d='M42.1198 33.9004C42.8101 33.9004 43.3698 33.3407 43.3698 32.6504C43.3698 31.96 42.8101 31.4004 42.1198 31.4004C41.4294 31.4004 40.8698 31.96 40.8698 32.6504C40.8698 33.3407 41.4294 33.9004 42.1198 33.9004Z'
-        fill='#23282E'
-      />
-      <path
-        d='M22.4397 47.6504C23.1301 47.6504 23.6897 47.0907 23.6897 46.4004C23.6897 45.71 23.1301 45.1504 22.4397 45.1504C21.7493 45.1504 21.1897 45.71 21.1897 46.4004C21.1897 47.0907 21.7493 47.6504 22.4397 47.6504Z'
-        fill='#23282E'
-      />
-      <path
-        d='M31.8097 47.6504C32.5 47.6504 33.0597 47.0907 33.0597 46.4004C33.0597 45.71 32.5 45.1504 31.8097 45.1504C31.1193 45.1504 30.5597 45.71 30.5597 46.4004C30.5597 47.0907 31.1193 47.6504 31.8097 47.6504Z'
-        fill='#23282E'
-      />
-      <path
-        d='M25.2498 48.9004H28.9998'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M36.8097 35.1504H40.5597'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <g opacity='0.3'>
-        <path
-          d='M58.41 18.3112C53.63 22.5712 62.49 23.2212 66.7 22.9712C65.45 18.7312 63.36 14.8512 60.6 11.5412C53.02 11.2712 46.04 17.1912 58.41 18.3112Z'
-          fill='white'
-        />
-        <path
-          d='M0.23999 28.1709C0.0899902 29.4309 0 30.7009 0 32.0009V35.9109C9 35.2409 9.82999 28.6709 0.23999 28.1809V28.1709Z'
-          fill='white'
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default WeatherTwo;
