--- components/icons/weather/WeatherZero.tsx
@@ -1,92 +1,0 @@
-import React from 'react';
-
-const WeatherZero = () => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='68'
-      height='100'
-      viewBox='0 0 68 100'
-      fill='none'
-    >
-      <path
-        d='M36 0H32C14.33 0 0 14.33 0 32V88C0 94.63 5.37 100 12 100H56C62.63 100 68 94.63 68 88V32C68 14.33 53.67 0 36 0Z'
-        fill='#B6F2FF'
-      />
-      <path
-        d='M53.44 79.6193C27.75 82.8893 30.44 66.7093 21.96 67.6893C18.71 68.0593 17.32 72.2793 15.69 69.1793C7.96 54.4893 3.32 59.1193 0 67.7393V89.0793H64.12V79.4693L53.44 79.6193Z'
-        fill='#55B58E'
-      />
-      <path
-        d='M41.95 73.2109C28.23 77.2909 14.51 82.1609 0 81.5009V88.0009C0 94.6309 5.37 100.001 12 100.001H56C62.63 100.001 68 94.6309 68 88.0009V70.3609C59.34 69.1809 50.42 70.8509 41.95 73.2109Z'
-        fill='#7AD7A9'
-      />
-      <g opacity='0.3'>
-        <path
-          d='M58.41 18.3102C53.63 22.5702 62.49 23.2202 66.7 22.9702C65.45 18.7302 63.36 14.8502 60.6 11.5402C53.02 11.2702 46.04 17.1902 58.41 18.3102Z'
-          fill='white'
-        />
-        <path
-          d='M0.24 28.1699C0.0899998 29.4299 0 30.6999 0 31.9999V35.9099C9 35.2399 9.83 28.6699 0.24 28.1799V28.1699Z'
-          fill='white'
-        />
-      </g>
-      <path
-        d='M33.9996 53.1211C43.6646 53.1211 51.4996 45.2861 51.4996 35.6211C51.4996 25.9561 43.6646 18.1211 33.9996 18.1211C24.3347 18.1211 16.4996 25.9561 16.4996 35.6211C16.4996 45.2861 24.3347 53.1211 33.9996 53.1211Z'
-        fill='#FF5C46'
-      />
-      <path
-        d='M33.9996 21.8711C26.4096 21.8711 20.2496 28.0311 20.2496 35.6211C20.2496 43.2111 26.4096 49.3711 33.9996 49.3711C41.5896 49.3711 47.7496 43.2111 47.7496 35.6211C47.7496 28.0311 41.5896 21.8711 33.9996 21.8711ZM33.9996 45.0011C28.8196 45.0011 24.6196 40.8011 24.6196 35.6211C24.6196 30.4411 28.8196 26.2411 33.9996 26.2411C39.1796 26.2411 43.3796 30.4411 43.3796 35.6211C43.3796 40.8011 39.1796 45.0011 33.9996 45.0011Z'
-        fill='#FFD34C'
-      />
-      <path
-        d='M33.9996 26.251C28.8196 26.251 24.6196 30.451 24.6196 35.631C24.6196 40.811 28.8196 45.011 33.9996 45.011C39.1796 45.011 43.3796 40.811 43.3796 35.631C43.3796 30.451 39.1796 26.251 33.9996 26.251ZM33.9996 40.621C31.2396 40.621 28.9996 38.381 28.9996 35.621C28.9996 32.861 31.2396 30.621 33.9996 30.621C36.7596 30.621 38.9996 32.861 38.9996 35.621C38.9996 38.381 36.7596 40.621 33.9996 40.621Z'
-        fill='#7AD7A9'
-      />
-      <path
-        d='M33.9996 40.6211C36.7611 40.6211 38.9996 38.3825 38.9996 35.6211C38.9996 32.8597 36.7611 30.6211 33.9996 30.6211C31.2382 30.6211 28.9996 32.8597 28.9996 35.6211C28.9996 38.3825 31.2382 40.6211 33.9996 40.6211Z'
-        fill='#5A7EFF'
-      />
-      <path
-        d='M44.8097 38.1607C43.6597 38.0707 42.5397 38.1707 41.4897 38.4507C40.0197 38.8307 38.4797 38.2607 37.5597 37.0607C35.0897 33.8207 31.1497 31.7707 26.7397 31.8907C19.8897 32.0807 14.2597 37.6407 13.9997 44.5007C13.7197 51.9807 19.6997 58.1407 27.1197 58.1407H43.9997C49.7497 58.1407 54.3697 53.2907 53.9797 47.4507C53.6497 42.5507 49.7097 38.5607 44.8197 38.1707L44.8097 38.1607Z'
-        fill='white'
-      />
-      <path
-        d='M25.2496 47.501C26.4996 48.441 27.7496 48.441 28.9996 47.501'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeMiterlimit='10'
-        strokeLinecap='round'
-      />
-      <path
-        d='M21.3297 43.1211L23.3697 45.0011L21.3297 46.8711'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M20.5597 45.001H23.3697'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M32.9196 43.1211L30.8696 45.0011L32.9196 46.8711'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-      <path
-        d='M33.6896 45.001H30.8696'
-        stroke='#23282E'
-        strokeWidth='1.25'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  );
-};
-export default WeatherZero;
